The javascript will control the event preview.
By default all events will show in the immediate events section. If
Javascript is enabled then we will remove expired events and hide all
but the next two events (which can be disclosed).

This wasn't working as we were doing part of this in JS and part in
Jekyll - which clearly was dumb. Oh well.